refactor to avoid circular imports
