+++ matchms/__init__.py
@@ -7,9 +7,8 @@ from .__version__ import __version__
 from .calculate_scores import calculate_scores
 from .logging_functions import _init_logger
 from .logging_functions import set_matchms_logger_level
-#from .Metadata import Metadata
+from .Metadata import Metadata
 from .Scores import Scores
-from .Spectrum import Metadata
 from .Spectrum import Spectrum
 from .Fragments import Fragments
