+++ matchms/filtering/derive_ionmode.py
@@ -1,5 +1,5 @@
 import logging
-from ..importing import load_adducts_dict
+from matchms.filtering.load_adducts import load_adducts_dict
 from ..typing import SpectrumType
 from ..metadata_utils import clean_adduct
