+++ matchms/importing/__init__.py
@@ -14,8 +14,6 @@ For more extensive import options we recommend building custom importers using `
 To process spectrum metadata, matchms can also make use of known adduct information
 which is imported via :mod:`~matchms.importing.load_adducts`.
 """
-from .load_adducts import load_adducts_dict
-from .load_adducts import load_known_adduct_conversions
 from .load_from_json import load_from_json
 from .load_from_mgf import load_from_mgf
 from .load_from_msp import load_from_msp
@@ -31,6 +29,4 @@ __all__ = [
     "load_from_mzml",
     "load_from_mzxml",
     "load_from_usi",
-    "load_adducts_dict",
-    "load_known_adduct_conversions",
 ]
